backend/process-image/requirements.txt
Pillow==10.1.0
requests==2.31.0
numpy==1.24.3